# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FacilityTrack Aquatic Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /steam-room-checks:
    get:
      operationId: listSteamRoomChecks
      tags: [steam-room]
      summary: List steam room checks
      parameters:
        - name: poolId
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of steam room checks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SteamRoomCheck"
    post:
      operationId: createSteamRoomCheck
      tags: [steam-room]
      summary: Log a steam room check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSteamRoomCheckBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SteamRoomCheck"
  /steam-room-checks/{id}:
    get:
      operationId: getSteamRoomCheck
      tags: [steam-room]
      summary: Get a steam room check
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Steam room check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SteamRoomCheck"
        "404":
          description: Not found
    patch:
      operationId: updateSteamRoomCheck
      tags: [steam-room]
      summary: Update a steam room check
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSteamRoomCheckBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SteamRoomCheck"
    delete:
      operationId: deleteSteamRoomCheck
      tags: [steam-room]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── WORK ORDERS ──────────────────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: facilityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isActive
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          description: Not found
    patch:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── STAFF QUALIFICATIONS ─────────────────────────────────────────────────────
  /staff-qualifications:
    get:
      operationId: listStaffQualifications
      tags: [staff-qualifications]
      summary: List staff qualifications
      parameters:
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of qualifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffQualification"
    post:
      operationId: createStaffQualification
      tags: [staff-qualifications]
      summary: Add a staff qualification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffQualificationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffQualification"
  /staff-qualifications/{id}:
    get:
      operationId: getStaffQualification
      tags: [staff-qualifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Qualification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffQualification"
        "404":
          description: Not found
    patch:
      operationId: updateStaffQualification
      tags: [staff-qualifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffQualificationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffQualification"
    delete:
      operationId: deleteStaffQualification
      tags: [staff-qualifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── TRAINING RECORDS ─────────────────────────────────────────────────────────
  /training-records:
    get:
      operationId: listTrainingRecords
      tags: [training-records]
      summary: List training records
      parameters:
        - name: staffId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of training records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingRecord"
    post:
      operationId: createTrainingRecord
      tags: [training-records]
      summary: Create a training record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingRecordBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"
  /training-records/{id}:
    get:
      operationId: getTrainingRecord
      tags: [training-records]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"
        "404":
          description: Not found
    patch:
      operationId: updateTrainingRecord
      tags: [training-records]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainingRecordBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecord"
    delete:
      operationId: deleteTrainingRecord
      tags: [training-records]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SteamRoomCheck:
      type: object
      properties:
        id:
          type: integer
        poolId:
          type: integer
        poolName:
          type: ["string", "null"]
        checkedBy:
          type: ["string", "null"]
        checkedAt:
          type: string
          format: date-time
        temperature:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        isClean:
          type: ["boolean", "null"]
        isOperational:
          type: ["boolean", "null"]
        entryType:
          type: string
          enum: [day_log, tablet_entry]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - poolId
        - checkedAt
        - entryType
        - createdAt
    CreateSteamRoomCheckBody:
      type: object
      properties:
        poolId:
          type: integer
        checkedBy:
          type: ["string", "null"]
        checkedAt:
          type: ["string", "null"]
        temperature:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        isClean:
          type: ["boolean", "null"]
        isOperational:
          type: ["boolean", "null"]
        entryType:
          type: string
          enum: [day_log, tablet_entry]
        notes:
          type: ["string", "null"]
      required:
        - poolId
        - entryType
    UpdateSteamRoomCheckBody:
      type: object
      properties:
        temperature:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        isClean:
          type: ["boolean", "null"]
        isOperational:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]

    # ── WORK ORDER ────────────────────────────────────────────────────────────
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        facilityId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - isActive
        - createdAt
        - updatedAt
    CreateStaffBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        facilityId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        startDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - firstName
        - lastName
    UpdateStaffBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        startDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── STAFF QUALIFICATION ───────────────────────────────────────────────────
    StaffQualification:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        qualificationName:
          type: string
        issuer:
          type: ["string", "null"]
        issuedDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        certificateNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [current, expiring_soon, expired]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - staffId
        - qualificationName
        - status
        - createdAt
    CreateStaffQualificationBody:
      type: object
      properties:
        staffId:
          type: integer
        qualificationName:
          type: string
        issuer:
          type: ["string", "null"]
        issuedDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        certificateNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - staffId
        - qualificationName
    UpdateStaffQualificationBody:
      type: object
      properties:
        qualificationName:
          type: ["string", "null"]
        issuer:
          type: ["string", "null"]
        issuedDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        certificateNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── TRAINING RECORD ───────────────────────────────────────────────────────
    TrainingRecord:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        trainingName:
          type: string
        completedAt:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        durationHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - staffId
        - trainingName
        - createdAt
    CreateTrainingRecordBody:
      type: object
      properties:
        staffId:
          type: integer
        trainingName:
          type: string
        completedAt:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        durationHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - staffId
        - trainingName
    UpdateTrainingRecordBody:
      type: object
      properties:
        trainingName:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        provider:
          type: ["string", "null"]
        durationHours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ── NOTIFICATION ──────────────────────────────────────────────────────────
...